Avoid get_initial_value when we have a tool state
It's not guaranteed this will always work, and it's unnecessary if
we have e.g. `ConnectedValue` in a workflow context.
